refactor: make animation reusable
